Create test helpers: less repetition, more readability
